Fixed warning C4018: '<': signed/unsigned mismatch
The right-hand side `bytes_read` is a `DWORD`, which is an `unsigned
long` on Windows.

+++ src/platform_win.cc
@@ -322,7 +322,7 @@ optional<std::string> RunExecutable(const std::vector<std::string>& command,
       if (!success || bytes_read == 0)
         break;
 
-      for (int i = 0; i < bytes_read; ++i)
+      for (unsigned long i = 0; i < bytes_read; ++i)
         output += buffer[i];
     }
   };
